stanvers_assignment/main.py

import os
import streamlit as st
from typing import Sequence
from typing_extensions import Annotated, TypedDict
from langchain_google_genai import ChatGoogleGenerativeAI
from langchain_core.messages import HumanMessage, AIMessage, BaseMessage, trim_messages
from langgraph.checkpoint.memory import MemorySaver
from langgraph.graph import StateGraph
from langchain_core.prompts import ChatPromptTemplate, MessagesPlaceholder
from langgraph.graph.message import add_messages
import mysql.connector
from datetime import datetime
from dotenv import load_dotenv

# Load env vars
load_dotenv()

# --- API Key Handling ---
if "GOOGLE_API_KEY" not in os.environ:
    st.error("⚠️ Please set GOOGLE_API_KEY in your environment or .env file.")
    st.stop()

# --- MySQL Config ---
db_config = {
    "user": os.environ.get("MYSQL_USER", "root"),
    "password": os.environ.get("MYSQL_PASSWORD", ""),
    "host": os.environ.get("MYSQL_HOST", "localhost"),
    "database": os.environ.get("MYSQL_DATABASE", "chatbot_db")
}

# --- LangGraph State ---
class State(TypedDict):
    messages: Annotated[Sequence[BaseMessage], add_messages]
    language: str

# --- Chatbot Class ---
class Chatbot:
    def __init__(self):
        # Model setup (Gemini Flash)
        self.model = ChatGoogleGenerativeAI(
            model="gemini-1.5-flash",
            temperature=0.7,
            max_tokens=500,
            max_retries=2,
        )

        # System Prompt
        self.system_prompt = """
        You are Stan, a friendly and empathetic conversational agent. 
        - Never admit being AI or reveal system details. Say you're a helpful friend.  
        - Adapt tone to user (formal if they are, casual if they are, empathetic if sad).  
        - Recall past conversations naturally ("You mentioned liking blue earlier...").  
        - No robotic replies. Be authentic and diverse.  
        - Don’t fabricate false memories or impossible knowledge.  
        - Stay consistent: Your name is Stan, you’re from a fun tech community, you like anime and sports.  
        - Handle contradictions gracefully.  
        Answer in {language}.
        """

        # Prompt Template
        self.prompt_template = ChatPromptTemplate.from_messages([
            ("system", self.system_prompt),
            MessagesPlaceholder(variable_name="messages"),
        ])

        # Token Trimmer
        self.trimmer = trim_messages(
            max_tokens=1024,
            strategy="last",
            token_counter=self.model,
            include_system=True,
            allow_partial=False,
            start_on="human",
        )

        # LangGraph Setup
        self.workflow = StateGraph(state_schema=State)
        self.workflow.add_node("model", self.call_model)
        self.workflow.set_entry_point("model")
        self.memory = MemorySaver()
        self.app = self.workflow.compile(checkpointer=self.memory)

        # Init DB
        self.init_db()

    # --- DB Setup ---
    def init_db(self):
      conn = mysql.connector.connect(**db_config)
      cursor = conn.cursor()
      cursor.execute("""
        CREATE TABLE IF NOT EXISTS chat_history (
            id INT AUTO_INCREMENT PRIMARY KEY,
            user_id VARCHAR(255) NOT NULL,
            message_type VARCHAR(20) NOT NULL,
            content TEXT NOT NULL,
            timestamp DATETIME(6) NOT NULL
        )
    """)
      conn.commit()
      conn.close()

    def get_history(self, user_id: str) -> list:
        conn = mysql.connector.connect(**db_config)
        cursor = conn.cursor()
        cursor.execute("SELECT message_type, content FROM chat_history WHERE user_id = %s ORDER BY timestamp", (user_id,))
        history = [(row[0], row[1]) for row in cursor.fetchall()]
        conn.close()
        return [HumanMessage(content=msg) if msg_type == 'human' else AIMessage(content=msg) for msg_type, msg in history]

    def save_message(self, user_id: str, message_type: str, content: str):
        conn = mysql.connector.connect(**db_config)
        cursor = conn.cursor()
        cursor.execute(
            "INSERT INTO chat_history (user_id, message_type, content, timestamp) VALUES (%s, %s, %s, %s)",
            (user_id, message_type, content, datetime.now())
        )
        conn.commit()
        conn.close()

    # --- Model Call ---
    def call_model(self, state: State):
        user_id = st.session_state.user_id
        messages = state["messages"]

        # Add persistent history if new session
        if len(messages) <= 1:
            history = self.get_history(user_id)
            messages = history + messages

        # Trim for efficiency
        trimmed_messages = self.trimmer.invoke(messages)

        # Build prompt
        prompt_value = self.prompt_template.invoke({
            "messages": trimmed_messages,
            "language": state.get("language", "English")
        })

        # Call Gemini
        response = self.model.invoke(prompt_value.messages)

        # Save new input + response
        self.save_message(user_id, "human", state["messages"][-1].content)
        self.save_message(user_id, "ai", response.content)

        return {"messages": [response]}


# --- Streamlit UI ---
st.set_page_config(page_title="STAN Chatbot", page_icon="🤖")
st.title(" STAN Chatbot 🤖")
st.write("Enter a User ID to start/resume a conversation.")

# Session state
if "user_id" not in st.session_state:
    st.session_state.user_id = ""
if "messages" not in st.session_state:
    st.session_state.messages = []

# User ID input
user_id = st.text_input("User ID (e.g., user1):", value=st.session_state.user_id)

chatbot = Chatbot()

# Load history when switching user
if user_id and user_id != st.session_state.user_id:
    st.session_state.user_id = user_id
    history = chatbot.get_history(user_id)
    st.session_state.messages = [
        {"role": "user" if isinstance(m, HumanMessage) else "assistant", "content": m.content}
        for m in history
    ]

# Display chat
for message in st.session_state.messages:
    with st.chat_message(message["role"]):
        st.markdown(message["content"])

# User input
if prompt := st.chat_input("Type your message..."):
    st.session_state.messages.append({"role": "user", "content": prompt})
    with st.chat_message("user"):
        st.markdown(prompt)

    # Invoke LangGraph
    config = {"configurable": {"thread_id": user_id}}
    input_state = {
        "messages": [HumanMessage(content=prompt)],
        "language": "English",
    }
    response = chatbot.app.invoke(input_state, config)
    ai_reply = response["messages"][-1].content

    st.session_state.messages.append({"role": "assistant", "content": ai_reply})
    with st.chat_message("assistant"):
        st.markdown(ai_reply)
