stanvers_assignment/requirements.txt
streamlit
mysql-connector-python
python-dotenv
langchain
langchain-core
langchain-google-genai
langgraph